grafana-watcher: retry failed updates from ConfigMap changes
Currently, an update will only be partially applied if an
error occurs in one of the handler routines. The grafana
system will be left in a dirty intermediate state until
the ConfigMap is updated again. This change will retry failed updates
until all scheduled updates have succeeded OR until the ConfigMap
is changed again.