helm/prometheus, helm/alertmanager: add support for pod anti-affinity for Alertmanager and Prometheus (#889)
When running multiple instances of either Alertmanager or Prometheus it is better from a high availability point of view that the pods are not co-located. This commit introduces the usage of pod anti-affinity with a default setting of “soft” for both Alertmanager and Prometheus meaning the scheduler should make a best-effort to place replica pods on different nodes. The setting “hard” makes it required and “” disables the usage of pod anti-affinity.