crd: Enrich Prometheus operator CRD registration error handling
If a user:
1. deploys the Prometheus operator without CRD validation
2. creates an invalid CRD instance (e.g. Servicemonitor)
3. restarts the Prometheus operator
it will crash-loop with a difficult to understand error message, failing
to parse the invalid CRD instance.

This patch improves the error feedback to the user, to identify the
invalid CRD more easily. The Prometheus operator will keep
crash-looping to enforce a valid CRD state on startup.