pkg/version: Pass Prometheus Operator version via ldflags
The Prometheus config reloader image is released along with the
Prometheus Operator image, tagged with the same semver version. In order
to default to the Prometheus Operator version if no Prometheus config
reloader image is specified, the content of `VERSION` needs to be passed
during compile time.

In addition with both the Prometheus Operator and the Prometheus Config
Reloader being aware of their version, we can log it at process startup
time.