test/e2e: Fix wrong override of Prometheus config reloader arg
The e2e framework overrides the Prometheus config reloader argument of
the Prometheus Operator. Instead of overriding the correct argument, it
had been overriding the config reloader argument, resulting in two
specifications of the Prometheus config reloader argument one shadowing
the other.

This has only now caused problems as previously v0.21.0 would win, which
is present on quay.io. With the new release (v0.22.0) it fails as the
Prometheus config reloader v0.22.0 is not yet present on quay.io.

This patch resolves the wrong override and thereby fixes the e2e tests.