prometheus: Do not migrate rule config map for statefulset (#1679)
If a user sets the RuleSelector property of a Prometheus custom resource
to an empty object (`{}`), all config maps in that namespace will be
selected for migration, including the Prometheus config map used to
mount all PrometheusRules into the Prometheus statefulset. This patch
filters the above mentioned config map out.