client-go/inforers: Add namespace indexer to SharedIndexInformers
`client-go/tools/cache.ListAllByNamespace` tries to search the given
cache by the namespace index. If that index is not present, it logs a
warning message and iterates over the objects itself [1].

Initially to remove the log line, but as a side effect also to improve query
speed, this patch adds the namespace index to all used SharedIndexInformers in
the Prometheus and Alertmanager operator.

[1]
```
items, err := indexer.Index(NamespaceIndex, &metav1.ObjectMeta{Namespace: namespace})
if err != nil {
	// Ignore error; do slow search without index.
	glog.Warningf("can not retrieve list of objects using index : %v", err)
	for _, m := range indexer.List() {
```