pkg/*: sync caches before adding handlers
This commit ensures that all informer caches are synced before the
eventhandlers are added. This is important whenever one informer's
eventhandler depends on the cache of a different informer. Without this,
eventhandlers will complain that certain objects don't exist. This is
important now that the namespace informer still runs for a single
namespace, where previously it was only run for v1.NamespaceAll.
Otherwise, e.g. the secret eventhandler could fail until the namespace
handler had synced.