kube-prometheus: Ensure all images have an entry in _config.versions
I'd like to programatically inspect the imageRepos and versions maps to
generically manipulate the list of images used by the prometheus operator.

To do so, an interesting property is that every image in _config.imageRepos has
a corresponding entry in _config.versions.

The only one missing was prometheusConfigReloader, fortunately we can still
keep it DRY by re-using self.prometheusOperator.