Update the getting-started documentation ClusterRole definition
The ClusterRole scope on the definition provided by the manifest in the
getting started was too wide regarding customresourcedefinitions. Since
the resourceNames were not specified, the ClusteRole could modify any
customresourcedefinitions in the cluster.

This commit fixes that by reducing the scope of the permissions
regarding customresourcedefinitions, only to servicemonitors,
prometheusrules, prometheuses and alertmanagers.