Aggregate prometheus-crd-view role to admin
In case the `view` is not aggregated to `admin`, you would need to grant a user both `admin` and `view` roles on the same namespace to be able to edit and get the resources respectively. This means that an `admin` on the namespace would be able to `create` the resource object, but not `get` it. 

This change aggregates the view to the admin role, so that the admin can now `get` on the resource as well.