Add prometheusExternalLabelName field
This commit adds prometheusExternalLabelName field to Prometheus object.

The string value of this field is used to customize the name of
Prometheus external label holding namespace and name of Prometheus
object.

When the field is set to empty string (`""`) it would not be included
into a rendered external_labels dictionary of prometheus.yaml.

When the field is missing (`null`) it defaults to `prometheus` for
backward compatibility.