Update meaning of empty replicaExternalLabelName
Update how empty string value of replicaExternalLabelName in
PrometheusSpec is interpreted.

Prior to this commit, an empty string would result in a default value of
`prometheus_replica` being used, just as if the field was omitted.

In order to harmonize behavior with the way how we handle
prometheusExternalLabelName, we now enable users to suppress adding
global label entirely by setting replicaExternalLabelName to an empty
string.