feat: Change governing service port names to match istio standard
Istio requires that service port names follow a convention, and in particular that if two services point at the same pod (e.g. if a user makes a service with a ClusterIP in addition to the governing service) that the name's have the same protocol. By not changing the pod port name this preserves backwards compatibility with existing users services