Add ConfigSecret element to the AlertmanagerSpec.
Issue #2003 describes an issue with Kustomize naming of secrets and the
name used for discovering the alertmanager configuration secret. In a
nutshell, a secret created with kustomize will include the hash of the
secret values as a suffix of the object name which is unexpected by
alertmanager. Instead; when an alertmanager is created, its pods will
wait for a secret of the appropriate name to be created, when in reality
they will will not be.

This change adds a ConfigSecret element to the AlertmanagerSpec that
allows the name of the secret to be injected into the stateful set.

This change is strongly influenced by PR #2024; by @lswith.

Addresses #2003.