add operator for thanos-ruler
This change adds a new kubernetes CustomResourceDefinition and associated
operator/controller for ThanosRuler instances.
This allows ThanosRuler instances to be deployed according to the
configuration in the ThanosRuler objects.  A rule selector can be used
to select PrometheusRule objects which are combined into a ConfigMap and
mounted into the associated ThanosRuler pod.

Co-authored-by: Matthias Loibl <mail@matthiasloibl.com>
Co-authored-by: Sergiusz Urbaniak <sergiusz.urbaniak@gmail.com>