refactor update handling of prometheus statefulsets
Add a new hash annotation that tracks the state of the
statefulset spec separately from the inputs (Prometheus, Config,
ConfigMaps).  This hash annotation is added immediately
after the statefulset is created, and is checked for changes
during updates to detect if there were manual updates to the
statefulset spec.

This prevents an issue (#2950) where the statefulset is
continously deleted and then recreated in kubernetes v1.17
due to a mismatch between the hash annotation and the state
of the statefulsetspec.  The issue only occurs in kubernetes
v1.17 because the api is more strict about what parts of a
statefulset can be updated.