fix service monitor tls-file tests
The allowed-tls-file and denied-tls-file tests of service monitor config
were not able to access tls certificate files on the local filesystem.
This issue was only exposed when updating the default Prometheus version
above 2.8.0 because this version introduced validation of the TLS
config.

This changes the config for these two tests to mount a secret containing
these two files.