fix alertmanager statefulset being pruned by kubectl
kubectl-managed annotations will cause the generated statefulset to be pruned
when the source alertmanager is applied with kubectl apply --prune.

To fix this do not transfer any kubectl.kubernetes.io/ annotations to
the generated statefulset.

This is the same problem as fixed in #2944 but for alertmanager rather than
prometheus.

Fixes #3029

Signed-off-by: Alex Conlin-Oakley <alexconlin@gmail.com>