thanos-ruler 'ruledir' to glob rather than hardcoded path
Instead of hardcoding each name of each configmap of the rules that are generated, instead use a glob on the configmap name. This means we can have 0 or more instances of rules existing, and won't make the ThanosRuler StatefulSet object force a restart when a large number of rules are created.