thanos ruler statefulset comparison on rule update
When a PrometheusRule object is created the Thanos Ruler statefulset object is touched for update. The previous behaviour updated the statefulset regardless if it actually needed an update. This update creates a 'diff' check to see whether or not the statefulset needs to be updated, basically the same way that it is defined in the prometheus/operator.

We also force a hash key to be generated on the first go.