pkg/prometheus: deprecate pod monitor targetPort
The split between port and targetPort makes sense for ServiceMonitors
because this determines if we filter endpoints based on
endpoint_port_name or pod_container_port_name. However, for PodMonitors
this split doesn't make sense because there is no endpoint_port_name. So
we only need a single port config field.

Fixes #3071