pkg/thanos: fix thanos statefulset sync updates
Changes to the ThanosRuler objects were being ignored by the operator.
This was due to copying the hash annotation of the old statefulset
to the new one before doing the comparison, thus they were always
appearing to be equal.

There doesn't seem to be any reason why we are copying the annotations
from the old statefulset to the new one, so this has been removed.

Fixes #3081