Support matching only pod monitors
In an environment that only uses pod monitors, it is natural to only
specify a matcher for pod monitors. Support this by treating them just
like service monitors are treated.

The condition in `sync` was getting hairy. Move it down into
`createOrUpdateConfigurationSecret`, so that we can keep the happy-path
left-aligned overall.

This preserves the special rule that "no matchers" means "do not manage
the configuration at all" – this can be removed at a later stage. In the
meantime, improve the messaging around this in logs and documentation.

Add additional logging and documentation around this, including the
deprecation check.

Fixes #3164.

Signed-off-by: Matthias Rampke <mr@soundcloud.com>