reload alertmanager on configmap/secret change
Via the configMaps and secrets fields we can add additional
files that should be mounted into the /etc/alertmanager directory.
These can then be used for templates (e.g. e-mail templates).

If those change we want to reload the alertmanager in order
to urge it to take the changed files into account.

Signed-off-by: Tim Reddehase <tim.reddehase@new-work.se>