test/framework: fix prom active targets polling
In order to check if prometheus is able to scrape a target, we are
querying the /api/v1/targets endpoint and checking if the number of
active targets is equal to the one we expect. However, active targets
displayed by the API can sometimes be unhealthy. For example, when the
scraped target responds 401 Unauthorized, the target is still considered
active but its health will be set to "down". Because of that,
unauthorized scrapes will be considered valid in the tests.

Thus, in order to verify that prometheus is allowed to scrape a target,
we also need to check if it is healthy.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>