Fix expectation for TLS configs without local filesystem access
When local filesystem access is denied, values from secrets and config
maps can still be used. This is the desired behavior, as this is the way
to get secrets into Prometheus without access to its filesystem.

The tests gave the wrong result because validation of the TLS
configuration would always fail, not because it actually attempted to
access local files.

Signed-off-by: Matthias Rampke <matthias@rampke.de>