*: use a single reloader for Prometheus
This change gets rid of the additional reloader that was deployed in the
Prometheus pod to monitor rules file.

It also changes the Alertmanager and Thanos ruler statefulsets to use
the container image from the `--prometheus-config-reloader` flag instead
of `--config-reloader-image`. We pick `prometheus-config-reloader`
because it supports gzipped configuration file and is "smarter" about
when to reload the configuration (e.g. it checks that the content on
disk has effectively changed before reloading).

The `--config-reloader-image` of Prometheus operator is deprecated (e.g.
it isn't used anywhere by the operator) and it may be removed in a
future release (but not before v0.45).

Since the Prometheus config reloader doesn't watch only rule directories
anymore (e.g. for Alertmanager it watches mounted secrets), it gets a
new `--watched-dir` flag that replaces the `--rules-dir` flag.

Finally the config reloader can be configured to expose metrics with the
`--listen-address` flag. The metrics are:
* `reloader_config_apply_operations_failed_total`
* `reloader_config_apply_operations_total`
* `reloader_reloads_failed_total`
* `reloader_reloads_total`
* `reloader_watch_errors_total`
* `reloader_watches`
* `reloader_watch_events_total`

Signed-off-by: Simon Pasquier <spasquie@redhat.com>