po-docgen: fix doc generation for inlined objects
When po-docgen was generating the API documentation, it was
intentionally skipping inlined fields. Thus, they were absent from the
documentation instead of being displayed inline.
The assumption to skip inlined fields was correct for foreign types as
we most likely don't want their content to be part of the API
documentation. However, we want the content of internal types to be
inlined correctly.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>