*: fix reconciliation when 2 resources share the same secret
When the operator detects a change to a secret, service monitor, ..., it
enqueues the object's namespace and triggers a reconciliation for each
operated resource that watches this namespace. Instead of looping over
all operated resources, the operator stopped after the first resource.
If 2 Prometheus objects were depending on the same secret (referenced by
additionalScrapeConfigs for instance) then only one of them would be
reconciled when the secret was updated.

The regression has been introduced by #3440.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>