Remove pkg/errors dep from apis/monitoring module
I think this dependency is usually not needed any longer, given that
the standard library has the likes of errors.New, and fmt.Errorf.

My understanding is that apis/monitoring is a standalong module now to
make it easier for other projects to import it, with a minimal set of
transitive dependecies to worry about. If I'm correct, then
github.com/pkg/errors is an easy one to keep out.