*: fix child routes support in AlertmanagerConfig (#3703)
The `route` field of the AlertmanagerConfig CRD introduces a circular
reference which doesn't play well with the Kubernetes API. The end
result is that any sub-route from an AlertmanagerConfig object doesn't
show up in the generated configuration.

To get rid of the recursion, child routes are modeled as raw JSON
fields. The downside is that the API will only check that sub-routes
contain valid JSON data. The operator will ensure that the routes
definition is semantically correct.

Closes #3688

Signed-off-by: Simon Pasquier <spasquie@redhat.com>