Fix: Plural for Prometheus is Prometheis
After extensive research, it has been determined that the correct plural
for 'Prometheus' is 'Prometheis'. https://prometheus.io/docs/introduction/faq/#what-is-the-plural-of-prometheus

Let us use the correct plural, to avoid angering the Titan god of fire.

Since this is a breaking API change, VERSION was bumped accordingly. All
users of prometheus-operator will have to update their Charts. This is a small
sacrifice given the greater danger: One never knows, but the last thing
operators need is to have their holidays ruined by the wrath of Prometheus.

Happy Holidays everyone!