thanos-ruler: Always add web port to pod specs
When the Thanos Ruler CR is used together with Istio, we have to enable
the listenLocal option, so that the Thanos Ruler isn't bind against the
Pod IP.

When this option is enabled the "web" port isn't added to the Pod
specification, so that the corresponding endpoints for the Thanos Ruler
service doesn't contain the port. Because of this we can not use a
ServiceMonitor to scrape the Prometheus metrics of the Thanos Ruler in
this case.