*: reconcile on namespace updates (#3879)
Say that you have a Prometheus spec with
`serviceMonitorNamespaceSelector: {monitored: "true"}` and
`serviceMonitorSelector: {}` as well as a service monitor in namespace
"foo" which has the `monitored="true"` label. Then the operator wouldn't
remove the service monitor's targets when the `monitored="true"` label
is removed from namespace "foo" as one would expect. If namespace "foo"
doesn't have the selecting label initially, adding it later wouldn't
trigger the reconciliation either.

The same issue existed for pod monitors, probes, rules (Prometheus and
Thanos ruler) and alertmanager configurations.

This change adds logic to deal with namespace updates and trigger the
reconciliation if the namespace's labels have changed in a way that may
return different resources than previously. However it works only when
the operator uses a privileged lister/watcher on namespaces (e.g.  when
it isn't restricted to a set of allowed namespaces). In the other case,
the unprivileged lister/watcher doesn't propagate namespace updates as
expected, leaving the operator out-of-sync. This will be tackled in a
following change.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>