*: fix global HTTP configuration for Alertmanager
The bump of github.com/prometheus/common from v0.15.0 to v0.20.0
introduced a breaking change [1] for users that configure the
`http_config` parameter in the `global` section. In practice the
configuration generated by the operator would contain a
`follow_redirects: true` stanza that isn't supported by Alertmanager <=
v0.21.0.

Instead of importing the structs from github.com/prometheus/common, we
duplicate them in the internal representation of the Alertmanager
configuration.

[1] https://github.com/prometheus/common/commit/47ee79a16821e03c31ca741659d4f7919aed91c1

Signed-off-by: Simon Pasquier <spasquie@redhat.com>