pkg/operator: add timeout to informers cache sync
During heavy apiserver disruptions, the informers cache synchronization
done by the operator before running the informers might never return.
This result in the operator being stuck forever waiting for its
informers caches to be synchronized.
To prevent that, we introduce a timeout of 10 minutes in the context
passed to client-go to make sure that the sync will always return at
some point.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>