*: make log messages more consistent
When skipping fields in the generation of Alertmanager and Prometheus
configuration because they aren't supported by the current version,
ensure that we log the name and namespace of the resource as well as the
minimum required version and the current version.

This change also reduce code duplication in the generation of the
Prometheus configuration when enforcing sample, target and label limit.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>