pkg/prometheus: enforce namespace label always
Whenever `enforcedNamespaceLabel` is defined, the operator should add a
final relabeling configuration that sets the enforced label to the
namespace's value. It was effective for the `relabel_configs` field but
not for `metric_relabeling_configs`. The assumption was that dropping
metric relabeling entries that tried to alter the enforced label was
enough but it appears that the `labelmap` action could be used to
bypass the enforcement.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>