alertmanager: Support updated matching syntax for routes in CR
During the conversion of inhibit rules passed in the AlertManagerConfig
we prefer the updated matching syntax while maintaining support for deprecated
matchers.

The namespace label that gets injected as part of the conversion process is
set depending on the AM version.

Likewise, we fail both at conversion time (if a single CR is invalid it is thrown away),
and at config generation time (for raw config) if the input is not supported by the
AM version in the spec.