test: Reduce e2e test concurrency
A recent commit increased the concurrency for e2e tests and made it
double the amount of CPUs available on Github Runners. This seems to
have introduced some flakes in the e2e tests.

This commit reverts the concurrency setting and sets it to be the same
as the number of CPUs.