Don't propogate fluxcd labels
We use [fluxcd](https://github.com/fluxcd/flux2) with a
[kustomization](https://fluxcd.io/docs/components/kustomize/kustomization/)
referring to manifests generated by
[kube-prometheus](https://github.com/prometheus-operator/kube-prometheus)

Flux sets these labels on objects
it manages. If an object it manages exists on the cluster, but does not
exist in the defined kustomization, fluxcd will purge the object. As the
operator manages the statefulset, this does not exist in the
kustomization and flux will delete the statefulset when it reconciles.
We currently work around this by disabling flux garbage collection on
Prometheus objects, but it would be good if the operator didnt propogate
this label and we didnt need this workaround.