Extract common Prometheus fields in a separate struct
In preparation for adding the Prometheus Agent CRD, this commit extracts
fields common to the Prometheus server and agent into a dedicated struct
that can be embedded into the two CRDs.

The commit also refactors po-docgen and extracts the logic for parsing
structs and fields into a dedicate module. Furthermore, It improves
po-docgen to hide types which are only used as embedded, and merge
their fields into the types which extend them.