*: mark remoteWrite and remoteRead as non-experimental
The remote read and write features are established enough in Prometheus
that we can drop the "EXPERIMENTAL" warning in the Prometheus CRD and
support them officially from now on.

If the format of the remote write/read configuration in Prometheus
changes in a breaking way, the operator should deal with it based on the
`spec.version` field of the Prometheus resource, just like it does for
other configuration fields.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>