pkg/prometheus: reduce the number of reconciliations
Since the operator modifies the Prometheus status subresource, the
update event handler is called much more often than before which may
lead to higher CPU usage. To reduce the number of times the operator
runs the sync() method, it now compares the generation + labels +
annotations between the old and new objects (instead of using the
`resourceVersion` field) to decide whether a reconciliation is needed or
not. The same change is made for StatefulSet update handler with the
exception that the operator updates the Prometheus status subresource if
the `resourceVersion` value is different.

This is inspired by `GenerationChangedPredicate`,
`LabelChangedPredicate` and `AnnotationChangedPredicate` from
controller-runtime.

Closes #4784

Signed-off-by: Simon Pasquier <spasquie@redhat.com>