pkg/prometheus: add generation to computed hash
The `spec.resources` field of the Prometheus CRD embeds the
`resource.Quantity` struct which doesn't work with the hashstructure
library because some of its fields aren't public. The consequence is
that 2 Prometheus objects that differ only by their resources fields
could produce the same hash value and the operator wouldn't reconcile
the statefulset definition. To circumvent the issue, the object's
generation is added to the hash computation.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>